First merge fixes
