+++ app-next/app/page.js
@@ -1,9 +1,7 @@
 import HomePage from "@/components/HomePage/HomePage";
-import Navbar from "@/components/NavBar/Navbar";
 export default function Home() {
   return (
     <>
-      <Navbar />
       <HomePage />
     </>
   );
