+++ app-next/components/HomePage/HomePage.jsx
@@ -1,17 +1,16 @@
 import HYFLogo from "@/assets/hyf.svg";
 import Image from "next/image";
 import "./HomePage.css";
-import Hero from "../HeroSection/Hero";
-import WhyChooseSection from '../WhyChooseSection/WhyChooseSection.jsx';
-import PopularCoursesSection from '../PopularCoursesSection/PopularCoursesSection.jsx';
+import Hero from "../LandingPage/HeroSection/Hero.jsx";
+import WhyChooseSection from "../LandingPage/WhyChooseSection/WhyChooseSection.jsx";
+import PopularCoursesSection from "../LandingPage/PopularCoursesSection/PopularCoursesSection.jsx";
 
 export default function HomePage() {
-    return (
-        <div>
-            <Hero />
-            <WhyChooseSection />
-            <PopularCoursesSection />
-          </div>
-    );
+  return (
+    <div>
+      <Hero />
+      <WhyChooseSection />
+      <PopularCoursesSection />
+    </div>
+  );
 }
-
