--- app-next/components/PopularCoursesSection/PopularCoursesSection.module.css
@@ -1,102 +1,0 @@
-.container {
-    padding-top: 100px;
-    text-align: center;
-}
-
-.title {
-    font-family: Inter;
-    font-size: 36px;
-    font-weight: 700;
-    line-height: 40px;
-    color: #171A1F;
-    margin-bottom: 44px;
-}
-
-.coursesWrapper {
-    display: flex;
-    justify-content: center;
-    gap: 77px;
-    max-width: 1400px;
-    margin: 0 auto;
-}
-
-.courseCard {
-    width: 406px;
-    height: 518px;
-    background: #FFFFFF;
-    border-radius: 10px;
-    box-shadow: 0px 0px 1px #171a1f12, 0px 0px 2px #171a1f1F;
-    overflow: hidden;
-}
-
-.courseImage {
-    width: 100%;
-    height: 192px;
-}
-
-.courseContent {
-    padding: 24px;
-    text-align: left;
-}
-
-.courseTitle {
-    font-family: Inter;
-    font-size: 20px;
-    font-weight: 600;
-    line-height: 28px;
-    color: #171A1F;
-    margin-top: 10px;
-}
-
-.courseInstructor {
-    font-family: Roboto;
-    font-size: 18px;
-    font-weight: 700;
-    line-height: 28px;
-    color: #0072D6;
-    margin-bottom: 10px;
-}
-
-.courseDescription {
-    font-size: 16px;
-    color: #666;
-    margin-bottom: 20px;
-}
-
-.courseFooter {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin-top: 16px;
-}
-
-.coursePrice {
-    font-size: 18px;
-    font-weight: 700;
-    color: #171A1F;
-}
-
-.viewCourseButton {
-    padding: 12px 24px;
-    border: none;
-    border-radius: 8px;
-    background-color: #0072D6;
-    color: white;
-    font-size: 16px;
-    cursor: pointer;
-    font-weight: 600;
-}
-
-.browseAllButton {
-    margin-top: 176px;
-    margin-bottom: 80px;
-    padding: 16px 32px;
-    border: 1px solid #0072D6;
-    border-radius: 8px;
-    background-color: transparent;
-    color: #0072D6;
-    font-family: Roboto;
-    font-size: 18px;
-    font-weight: 500;
-    cursor: pointer;
-}
